# docs/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: A web of gin
  license: {}
  title: Golang Gin API
  version: "1.0"
paths:
...
  /api/v1/pay:
    post:
      parameters:
      - description: user_id
        in: body
        name: user_id
        schema:
          type: integer
      - description: order_id
        in: body
        name: order_id
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 微信下单 获取发起微信支付所需的数据
      tags:
      - 微信支付
  /api/v1/pay/taker:
    post:
      parameters:
      - description: user_id
        in: body
        name: user_id
        schema:
          type: integer
      - description: order_id
        in: body
        name: order_id
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 微信接单 获取发起微信支付所需的数据
      tags:
      - 微信支付
  /api/v1/phone/bind:
    post:
      parameters:
      - description: user_id
        in: body
        name: user_id
        schema:
          type: integer
      - description: phone
        in: body
        name: phone
        schema:
          type: string
      - description: code
        in: body
        name: code
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 绑定手机号 确认绑定的接口
      tags:
      - 绑定手机号
  /api/v1/phone/code:
    get:
      consumes:
      - application/json; charset=utf-8
      parameters:
      - description: user_id
        in: body
        name: user_id
        schema:
          type: integer
      - description: phone
        in: body
        name: phone
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取要绑定的手机号的验证码 绑定手机时使用
      tags:
      - 绑定手机号
  /api/v1/phone/code2:
    get:
      consumes:
      - application/json; charset=utf-8
      parameters:
      - description: user_id
        in: body
        name: user_id
        schema:
          type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取已绑定手机号的验证码 （不需要传手机号 验证提款密码时使用）
      tags:
      - 手机验证码
...
swagger: "2.0"